requirements.txt
# local package
-e .

# external requirements
#click
coverage
#awscli==1.27.38
flake8
python-dotenv>=0.5.1
torchmetrics==0.11.1
scikit-learn==1.2.1
wandb==0.13.10
pandas==1.5.3
numpy==1.23.5
tqdm==4.64.1
matplotlib==3.6.3
hydra-core==1.3.2
gdown==4.6.4
pydrive
google-cloud-storage
google-cloud
netcal==1.3.5
# other
pytest
pre-commit
